HEIA-4321: fixed progression for the selected text
Previously progression for selected text was equal to the top offset progression,  so when we select multiple text without scrolling the view their progression comes out to be same.
After this change all the selected text are going to have unique/different progression as now it is offset of the select rect.